refactor: separate driver structs into files
